Add files via upload
create a web application using Java Servlets that handles both GET and POST methods. Let's break down the steps to achieve this:

Create the Servlet:
Create a Java Servlet that extends HttpServlet and overrides the doGet and doPost methods. These methods will handle the GET and POST requests, respectively.Map the Servlet in web.xml:
In the web.xml file (deployment descriptor), you need to map the URL pattern to your servlet. This tells the web server which servlet should handle the incoming requestsCreate an HTML page that contains a form to input the first and second name. Set the form's method attribute to "POST" to send the data as a POST request.Now, when a user submits the form on the index.html page, the data will be sent as a POST request to the "/name" URL, which will be handled by the NameServlet. You can process the form data inside the doPost method and perform any required actions.

This example demonstrates a simple implementation of handling both GET and POST methods in a Java Servlet web application. You can expand on this foundation to add more functionalities and improve the user experience.